+++ src/Phan/Daemon/Transport/StreamResponder.php
@@ -52,7 +52,7 @@ class StreamResponder implements Responder
 
     /**
      * @return void
-     * @throws RuntimeException if called twice
+     * @throws \RuntimeException if called twice
      */
     public function sendResponseAndClose(array $data)
     {